Revert "Revert "Revert "Search component inside header"""
