--- src/App.css
@@ -1,16 +1,0 @@
-#scrollBarDesign::-webkit-scrollbar-track {
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  border-radius: 10px;
-  background-color: #f5f5f5;
-}
-
-#scrollBarDesign::-webkit-scrollbar {
-  width: 5px;
-  background-color: #f5f5f5;
-}
-
-#scrollBarDesign::-webkit-scrollbar-thumb {
-  border-radius: 10px;
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  background-color: purple;
-}
