--- src/components/Header/data.ts
@@ -1,1148 +1,0 @@
-export const data = {
-    albums: {
-        "href": "https://api.spotify.com/v1/search?query=rama&type=album&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=0&limit=5",
-        "items": [
-            {
-                "album_type": "album",
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/1mBydYMVBECdDmMfE2sEUO"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/1mBydYMVBECdDmMfE2sEUO",
-                        "id": "1mBydYMVBECdDmMfE2sEUO",
-                        "name": "Sachin-Jigar",
-                        "type": "artist",
-                        "uri": "spotify:artist:1mBydYMVBECdDmMfE2sEUO"
-                    }
-                ],
-                "available_markets":  [],
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/album/1VgXTSQsIrRn150j5MJrxR"
-                },
-                "href": "https://api.spotify.com/v1/albums/1VgXTSQsIrRn150j5MJrxR",
-                "id": "1VgXTSQsIrRn150j5MJrxR",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab67616d0000b273cab73e6ac9bf4bf3c6931570",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://i.scdn.co/image/ab67616d00001e02cab73e6ac9bf4bf3c6931570",
-                        "width": 300
-                    },
-                    {
-                        "height": 64,
-                        "url": "https://i.scdn.co/image/ab67616d00004851cab73e6ac9bf4bf3c6931570",
-                        "width": 64
-                    }
-                ],
-                "name": "Ramaiya Vastavaiya (Original Motion Picture Soundtrack)",
-                "release_date": "2013-07-19",
-                "release_date_precision": "day",
-                "total_tracks": 8,
-                "type": "album",
-                "uri": "spotify:album:1VgXTSQsIrRn150j5MJrxR"
-            },
-            {
-                "album_type": "album",
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/2YZyLoL8N0Wb9xBt1NhZWg"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/2YZyLoL8N0Wb9xBt1NhZWg",
-                        "id": "2YZyLoL8N0Wb9xBt1NhZWg",
-                        "name": "Kendrick Lamar",
-                        "type": "artist",
-                        "uri": "spotify:artist:2YZyLoL8N0Wb9xBt1NhZWg"
-                    }
-                ],
-                "available_markets": [
-
-                ],
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/album/4eLPsYPBmXABThSJ821sqY"
-                },
-                "href": "https://api.spotify.com/v1/albums/4eLPsYPBmXABThSJ821sqY",
-                "id": "4eLPsYPBmXABThSJ821sqY",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab67616d0000b2738b52c6b9bc4e43d873869699",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://i.scdn.co/image/ab67616d00001e028b52c6b9bc4e43d873869699",
-                        "width": 300
-                    },
-                    {
-                        "height": 64,
-                        "url": "https://i.scdn.co/image/ab67616d000048518b52c6b9bc4e43d873869699",
-                        "width": 64
-                    }
-                ],
-                "name": "DAMN.",
-                "release_date": "2017-04-14",
-                "release_date_precision": "day",
-                "total_tracks": 14,
-                "type": "album",
-                "uri": "spotify:album:4eLPsYPBmXABThSJ821sqY"
-            },
-            {
-                "album_type": "single",
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/4zCH9qm4R2DADamUHMCa6O"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/4zCH9qm4R2DADamUHMCa6O",
-                        "id": "4zCH9qm4R2DADamUHMCa6O",
-                        "name": "Anirudh Ravichander",
-                        "type": "artist",
-                        "uri": "spotify:artist:4zCH9qm4R2DADamUHMCa6O"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/6CXEwIaXYfVJ84biCxqc9k"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/6CXEwIaXYfVJ84biCxqc9k",
-                        "id": "6CXEwIaXYfVJ84biCxqc9k",
-                        "name": "Vishal Dadlani",
-                        "type": "artist",
-                        "uri": "spotify:artist:6CXEwIaXYfVJ84biCxqc9k"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/19LIHDDSHBD5NyYHI3gpzB"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/19LIHDDSHBD5NyYHI3gpzB",
-                        "id": "19LIHDDSHBD5NyYHI3gpzB",
-                        "name": "Shilpa Rao",
-                        "type": "artist",
-                        "uri": "spotify:artist:19LIHDDSHBD5NyYHI3gpzB"
-                    }
-                ],
-                "available_markets": [],
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/album/7r7X6pKpUy8RAQsInpoCzN"
-                },
-                "href": "https://api.spotify.com/v1/albums/7r7X6pKpUy8RAQsInpoCzN",
-                "id": "7r7X6pKpUy8RAQsInpoCzN",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab67616d0000b2736f05ba9f714ab37d7e908d54",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://i.scdn.co/image/ab67616d00001e026f05ba9f714ab37d7e908d54",
-                        "width": 300
-                    },
-                    {
-                        "height": 64,
-                        "url": "https://i.scdn.co/image/ab67616d000048516f05ba9f714ab37d7e908d54",
-                        "width": 64
-                    }
-                ],
-                "name": "Not Ramaiya Vastavaiya (From \"Jawan\")",
-                "release_date": "2023-08-29",
-                "release_date_precision": "day",
-                "total_tracks": 1,
-                "type": "album",
-                "uri": "spotify:album:7r7X6pKpUy8RAQsInpoCzN"
-            },
-            {
-                "album_type": "album",
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/1I7FVmvisCtSFzmm87mbLR"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/1I7FVmvisCtSFzmm87mbLR",
-                        "id": "1I7FVmvisCtSFzmm87mbLR",
-                        "name": "El Búho",
-                        "type": "artist",
-                        "uri": "spotify:artist:1I7FVmvisCtSFzmm87mbLR"
-                    }
-                ],
-                "available_markets": [],
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/album/0v8v8HAtfJChO5sdRh5EAx"
-                },
-                "href": "https://api.spotify.com/v1/albums/0v8v8HAtfJChO5sdRh5EAx",
-                "id": "0v8v8HAtfJChO5sdRh5EAx",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab67616d0000b27312f65a6926fb65185a8a2a56",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://i.scdn.co/image/ab67616d00001e0212f65a6926fb65185a8a2a56",
-                        "width": 300
-                    },
-                    {
-                        "height": 64,
-                        "url": "https://i.scdn.co/image/ab67616d0000485112f65a6926fb65185a8a2a56",
-                        "width": 64
-                    }
-                ],
-                "name": "Ramas",
-                "release_date": "2020-03-27",
-                "release_date_precision": "day",
-                "total_tracks": 19,
-                "type": "album",
-                "uri": "spotify:album:0v8v8HAtfJChO5sdRh5EAx"
-            },
-            {
-                "album_type": "album",
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/2YZyLoL8N0Wb9xBt1NhZWg"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/2YZyLoL8N0Wb9xBt1NhZWg",
-                        "id": "2YZyLoL8N0Wb9xBt1NhZWg",
-                        "name": "Kendrick Lamar",
-                        "type": "artist",
-                        "uri": "spotify:artist:2YZyLoL8N0Wb9xBt1NhZWg"
-                    }
-                ],
-                "available_markets": [
-
-                ],
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/album/79ONNoS4M9tfIA1mYLBYVX"
-                },
-                "href": "https://api.spotify.com/v1/albums/79ONNoS4M9tfIA1mYLBYVX",
-                "id": "79ONNoS4M9tfIA1mYLBYVX",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab67616d0000b2732e02117d76426a08ac7c174f",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://i.scdn.co/image/ab67616d00001e022e02117d76426a08ac7c174f",
-                        "width": 300
-                    },
-                    {
-                        "height": 64,
-                        "url": "https://i.scdn.co/image/ab67616d000048512e02117d76426a08ac7c174f",
-                        "width": 64
-                    }
-                ],
-                "name": "Mr. Morale & The Big Steppers",
-                "release_date": "2022-05-13",
-                "release_date_precision": "day",
-                "total_tracks": 19,
-                "type": "album",
-                "uri": "spotify:album:79ONNoS4M9tfIA1mYLBYVX"
-            }
-        ],
-        "limit": 5,
-        "next": "https://api.spotify.com/v1/search?query=rama&type=album&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=5&limit=5",
-        "offset": 0,
-        "previous": null,
-        "total": 156
-    },
-    artists: {
-        "href": "https://api.spotify.com/v1/search?query=sonu+soo&type=artist&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=0&limit=5",
-        "items": [
-            {
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/artist/1RViT01v3KsUvaEQk136nY"
-                },
-                "followers": {
-                    "href": null,
-                    "total": 1340
-                },
-                "genres": [],
-                "href": "https://api.spotify.com/v1/artists/1RViT01v3KsUvaEQk136nY",
-                "id": "1RViT01v3KsUvaEQk136nY",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab67616d0000b273e76039d563900c0c22637741",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://i.scdn.co/image/ab67616d00001e02e76039d563900c0c22637741",
-                        "width": 300
-                    },
-                    {
-                        "height": 64,
-                        "url": "https://i.scdn.co/image/ab67616d00004851e76039d563900c0c22637741",
-                        "width": 64
-                    }
-                ],
-                "name": "Sonu Sood",
-                "popularity": 24,
-                "type": "artist",
-                "uri": "spotify:artist:1RViT01v3KsUvaEQk136nY"
-            },
-            {
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/artist/1dVygo6tRFXC8CSWURQJq2"
-                },
-                "followers": {
-                    "href": null,
-                    "total": 14776030
-                },
-                "genres": [
-                    "filmi",
-                    "modern bollywood"
-                ],
-                "href": "https://api.spotify.com/v1/artists/1dVygo6tRFXC8CSWURQJq2",
-                "id": "1dVygo6tRFXC8CSWURQJq2",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab6761610000e5ebbc959d7569618ec2af2210f5",
-                        "width": 640
-                    },
-                    {
-                        "height": 320,
-                        "url": "https://i.scdn.co/image/ab67616100005174bc959d7569618ec2af2210f5",
-                        "width": 320
-                    },
-                    {
-                        "height": 160,
-                        "url": "https://i.scdn.co/image/ab6761610000f178bc959d7569618ec2af2210f5",
-                        "width": 160
-                    }
-                ],
-                "name": "Sonu Nigam",
-                "popularity": 75,
-                "type": "artist",
-                "uri": "spotify:artist:1dVygo6tRFXC8CSWURQJq2"
-            },
-            {
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/artist/3KU3cOhGlK6k2f79Rp8kRx"
-                },
-                "followers": {
-                    "href": null,
-                    "total": 74
-                },
-                "genres": [],
-                "href": "https://api.spotify.com/v1/artists/3KU3cOhGlK6k2f79Rp8kRx",
-                "id": "3KU3cOhGlK6k2f79Rp8kRx",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab67616d0000b273a1aebcd688cba45727b6ac4a",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://i.scdn.co/image/ab67616d00001e02a1aebcd688cba45727b6ac4a",
-                        "width": 300
-                    },
-                    {
-                        "height": 64,
-                        "url": "https://i.scdn.co/image/ab67616d00004851a1aebcd688cba45727b6ac4a",
-                        "width": 64
-                    }
-                ],
-                "name": "Sonu Sood Kotputli",
-                "popularity": 2,
-                "type": "artist",
-                "uri": "spotify:artist:3KU3cOhGlK6k2f79Rp8kRx"
-            },
-            {
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/artist/0uZJ0ainoRd9nSZK7RL9zB"
-                },
-                "followers": {
-                    "href": null,
-                    "total": 14
-                },
-                "genres": [],
-                "href": "https://api.spotify.com/v1/artists/0uZJ0ainoRd9nSZK7RL9zB",
-                "id": "0uZJ0ainoRd9nSZK7RL9zB",
-                "images": [],
-                "name": "Sonu Sood",
-                "popularity": 0,
-                "type": "artist",
-                "uri": "spotify:artist:0uZJ0ainoRd9nSZK7RL9zB"
-            },
-            {
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/artist/1Pn6pKlgzxcH6iIRp08dQr"
-                },
-                "followers": {
-                    "href": null,
-                    "total": 177866
-                },
-                "genres": [
-                    "filmi",
-                    "modern bollywood"
-                ],
-                "href": "https://api.spotify.com/v1/artists/1Pn6pKlgzxcH6iIRp08dQr",
-                "id": "1Pn6pKlgzxcH6iIRp08dQr",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://i.scdn.co/image/ab6761610000e5eb81da094d57db47eee247c21a",
-                        "width": 640
-                    },
-                    {
-                        "height": 320,
-                        "url": "https://i.scdn.co/image/ab6761610000517481da094d57db47eee247c21a",
-                        "width": 320
-                    },
-                    {
-                        "height": 160,
-                        "url": "https://i.scdn.co/image/ab6761610000f17881da094d57db47eee247c21a",
-                        "width": 160
-                    }
-                ],
-                "name": "Sonu Kakkar",
-                "popularity": 51,
-                "type": "artist",
-                "uri": "spotify:artist:1Pn6pKlgzxcH6iIRp08dQr"
-            }
-        ],
-        "limit": 5,
-        "next": "https://api.spotify.com/v1/search?query=sonu+soo&type=artist&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=5&limit=5",
-        "offset": 0,
-        "previous": null,
-        "total": 356
-    },
-    playlists: {
-        "href": "https://api.spotify.com/v1/search?query=sonu+soo&type=playlist&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=0&limit=5",
-        "items": [
-            {
-                "collaborative": false,
-                "description": "",
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/playlist/67kYT1abdp2jBInBGNSBMi"
-                },
-                "href": "https://api.spotify.com/v1/playlists/67kYT1abdp2jBInBGNSBMi",
-                "id": "67kYT1abdp2jBInBGNSBMi",
-                "images": [
-                    {
-                        "height": 640,
-                        "url": "https://mosaic.scdn.co/640/ab67616d0000b27301da1830a32f9bae37838351ab67616d0000b2733367df290ba209eb6065c907ab67616d0000b2738b3514b7aebeefe5c10a4612ab67616d0000b273edc4afff2478f5d4ab60b71e",
-                        "width": 640
-                    },
-                    {
-                        "height": 300,
-                        "url": "https://mosaic.scdn.co/300/ab67616d0000b27301da1830a32f9bae37838351ab67616d0000b2733367df290ba209eb6065c907ab67616d0000b2738b3514b7aebeefe5c10a4612ab67616d0000b273edc4afff2478f5d4ab60b71e",
-                        "width": 300
-                    },
-                    {
-                        "height": 60,
-                        "url": "https://mosaic.scdn.co/60/ab67616d0000b27301da1830a32f9bae37838351ab67616d0000b2733367df290ba209eb6065c907ab67616d0000b2738b3514b7aebeefe5c10a4612ab67616d0000b273edc4afff2478f5d4ab60b71e",
-                        "width": 60
-                    }
-                ],
-                "name": "Soothing Sonu",
-                "owner": {
-                    "display_name": "RONY GHOSH DASTIDER",
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/user/316g2xoler4nh2wyhusoi6bwhjtm"
-                    },
-                    "href": "https://api.spotify.com/v1/users/316g2xoler4nh2wyhusoi6bwhjtm",
-                    "id": "316g2xoler4nh2wyhusoi6bwhjtm",
-                    "type": "user",
-                    "uri": "spotify:user:316g2xoler4nh2wyhusoi6bwhjtm"
-                },
-                "primary_color": null,
-                "public": null,
-                "snapshot_id": "MzYsM2I0NDQxMWUyYmQwMzBhYjliZjY4NTQwMjU0ZGNlOWVlMjk5NmZhZQ==",
-                "tracks": {
-                    "href": "https://api.spotify.com/v1/playlists/67kYT1abdp2jBInBGNSBMi/tracks",
-                    "total": 31
-                },
-                "type": "playlist",
-                "uri": "spotify:playlist:67kYT1abdp2jBInBGNSBMi"
-            },
-            {
-                "collaborative": false,
-                "description": "",
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/playlist/7hU8CylZielBw01H15Yxpx"
-                },
-                "href": "https://api.spotify.com/v1/playlists/7hU8CylZielBw01H15Yxpx",
-                "id": "7hU8CylZielBw01H15Yxpx",
-                "images": [
-                    {
-                        "height": null,
-                        "url": "https://image-cdn-ak.spotifycdn.com/image/ab67706c0000bebb8bd7ef1a6a7114c6874d4559",
-                        "width": null
-                    }
-                ],
-                "name": "Sonu Nigam Bengali Songs",
-                "owner": {
-                    "display_name": "__Time.For.Memories..",
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/user/21vmmy7ws2a6opbrsf2ns6ujy"
-                    },
-                    "href": "https://api.spotify.com/v1/users/21vmmy7ws2a6opbrsf2ns6ujy",
-                    "id": "21vmmy7ws2a6opbrsf2ns6ujy",
-                    "type": "user",
-                    "uri": "spotify:user:21vmmy7ws2a6opbrsf2ns6ujy"
-                },
-                "primary_color": null,
-                "public": null,
-                "snapshot_id": "NjQsNDBjZTY2MzZmNDNkOWY2ZDhjZDEwZjlmOTlhN2ZlOGQxM2E0ZDk3Mg==",
-                "tracks": {
-                    "href": "https://api.spotify.com/v1/playlists/7hU8CylZielBw01H15Yxpx/tracks",
-                    "total": 27
-                },
-                "type": "playlist",
-                "uri": "spotify:playlist:7hU8CylZielBw01H15Yxpx"
-            },
-            {
-                "collaborative": false,
-                "description": "sööyleee neeerdesiiin baaal",
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/playlist/0BJ8eutkZFHUcaPdcJ7yJr"
-                },
-                "href": "https://api.spotify.com/v1/playlists/0BJ8eutkZFHUcaPdcJ7yJr",
-                "id": "0BJ8eutkZFHUcaPdcJ7yJr",
-                "images": [
-                    {
-                        "height": null,
-                        "url": "https://image-cdn-ak.spotifycdn.com/image/ab67706c0000bebb89f6c9c4c952577ea3f01b08",
-                        "width": null
-                    }
-                ],
-                "name": "her güzel şeyin bir sonu vardır",
-                "owner": {
-                    "display_name": "İpek",
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/user/jjuey0wmb1pch2wef9zhj9dta"
-                    },
-                    "href": "https://api.spotify.com/v1/users/jjuey0wmb1pch2wef9zhj9dta",
-                    "id": "jjuey0wmb1pch2wef9zhj9dta",
-                    "type": "user",
-                    "uri": "spotify:user:jjuey0wmb1pch2wef9zhj9dta"
-                },
-                "primary_color": null,
-                "public": null,
-                "snapshot_id": "MjY4LDBkNjIzOTk0ODBmMmIxYjYyZGFkZmRiY2QwODg5NDhkNDgyYTNkMWY=",
-                "tracks": {
-                    "href": "https://api.spotify.com/v1/playlists/0BJ8eutkZFHUcaPdcJ7yJr/tracks",
-                    "total": 221
-                },
-                "type": "playlist",
-                "uri": "spotify:playlist:0BJ8eutkZFHUcaPdcJ7yJr"
-            },
-            {
-                "collaborative": false,
-                "description": "This is Sonu Nigam. The essential tracks, all in one playlist.",
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/playlist/37i9dQZF1DZ06evO0FcUGj"
-                },
-                "href": "https://api.spotify.com/v1/playlists/37i9dQZF1DZ06evO0FcUGj",
-                "id": "37i9dQZF1DZ06evO0FcUGj",
-                "images": [
-                    {
-                        "height": null,
-                        "url": "https://thisis-images.spotifycdn.com/37i9dQZF1DZ06evO0FcUGj-large.jpg",
-                        "width": null
-                    }
-                ],
-                "name": "This Is Sonu Nigam",
-                "owner": {
-                    "display_name": "Spotify",
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/user/spotify"
-                    },
-                    "href": "https://api.spotify.com/v1/users/spotify",
-                    "id": "spotify",
-                    "type": "user",
-                    "uri": "spotify:user:spotify"
-                },
-                "primary_color": null,
-                "public": null,
-                "snapshot_id": "Mjg0MjQ4NzgsMDAwMDAwMDA3ZDEwYjdmMGQ5YWM0MzA1ZDRmZTc2M2JhNDBlZjk4Yg==",
-                "tracks": {
-                    "href": "https://api.spotify.com/v1/playlists/37i9dQZF1DZ06evO0FcUGj/tracks",
-                    "total": 54
-                },
-                "type": "playlist",
-                "uri": "spotify:playlist:37i9dQZF1DZ06evO0FcUGj"
-            },
-            {
-                "collaborative": false,
-                "description": "GUYS ARIJIT IS SINGHING A SONG...... btw cover photo is of srk and sonu sood 😻 ",
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/playlist/7zRHuRCw3iQBjVyHsRtVFO"
-                },
-                "href": "https://api.spotify.com/v1/playlists/7zRHuRCw3iQBjVyHsRtVFO",
-                "id": "7zRHuRCw3iQBjVyHsRtVFO",
-                "images": [
-                    {
-                        "height": null,
-                        "url": "https://image-cdn-ak.spotifycdn.com/image/ab67706c0000bebb6fe73eec344712e181e9e004",
-                        "width": null
-                    }
-                ],
-                "name": "🌷",
-                "owner": {
-                    "display_name": "Sanjana's dead fish",
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/user/31rn5o5bzgzp4sv5vjflspyrr2im"
-                    },
-                    "href": "https://api.spotify.com/v1/users/31rn5o5bzgzp4sv5vjflspyrr2im",
-                    "id": "31rn5o5bzgzp4sv5vjflspyrr2im",
-                    "type": "user",
-                    "uri": "spotify:user:31rn5o5bzgzp4sv5vjflspyrr2im"
-                },
-                "primary_color": null,
-                "public": null,
-                "snapshot_id": "MTAyLGVkOTMwYjBkOWQxZGE4MWM2MDE1ZmMxZDQxMGU4MmViZDM4NWM3Zjc=",
-                "tracks": {
-                    "href": "https://api.spotify.com/v1/playlists/7zRHuRCw3iQBjVyHsRtVFO/tracks",
-                    "total": 36
-                },
-                "type": "playlist",
-                "uri": "spotify:playlist:7zRHuRCw3iQBjVyHsRtVFO"
-            }
-        ],
-        "limit": 5,
-        "next": "https://api.spotify.com/v1/search?query=sonu+soo&type=playlist&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=5&limit=5",
-        "offset": 0,
-        "previous": null,
-        "total": 815
-    },
-    tracks: {
-        "href": "https://api.spotify.com/v1/search?query=sonu+soo&type=track&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=0&limit=5",
-        "items": [
-            {
-                "album": {
-                    "album_type": "album",
-                    "artists": [
-                        {
-                            "external_urls": {
-                                "spotify": "https://open.spotify.com/artist/6hpDlOvqsnIxrPegGTWVY1"
-                            },
-                            "href": "https://api.spotify.com/v1/artists/6hpDlOvqsnIxrPegGTWVY1",
-                            "id": "6hpDlOvqsnIxrPegGTWVY1",
-                            "name": "Shantanu Moitra",
-                            "type": "artist",
-                            "uri": "spotify:artist:6hpDlOvqsnIxrPegGTWVY1"
-                        }
-                    ],
-                    "available_markets": [
-                    ],
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/album/2ybn771hbNo5dTTHrlZBmo"
-                    },
-                    "href": "https://api.spotify.com/v1/albums/2ybn771hbNo5dTTHrlZBmo",
-                    "id": "2ybn771hbNo5dTTHrlZBmo",
-                    "images": [
-                        {
-                            "height": 640,
-                            "url": "https://i.scdn.co/image/ab67616d0000b273db8153f339baece275dbe9d4",
-                            "width": 640
-                        },
-                        {
-                            "height": 300,
-                            "url": "https://i.scdn.co/image/ab67616d00001e02db8153f339baece275dbe9d4",
-                            "width": 300
-                        },
-                        {
-                            "height": 64,
-                            "url": "https://i.scdn.co/image/ab67616d00004851db8153f339baece275dbe9d4",
-                            "width": 64
-                        }
-                    ],
-                    "name": "Parineeta (Original Motion Picture Soundtrack)",
-                    "release_date": "2018-07-11",
-                    "release_date_precision": "day",
-                    "total_tracks": 7,
-                    "type": "album",
-                    "uri": "spotify:album:2ybn771hbNo5dTTHrlZBmo"
-                },
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/1dVygo6tRFXC8CSWURQJq2"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/1dVygo6tRFXC8CSWURQJq2",
-                        "id": "1dVygo6tRFXC8CSWURQJq2",
-                        "name": "Sonu Nigam",
-                        "type": "artist",
-                        "uri": "spotify:artist:1dVygo6tRFXC8CSWURQJq2"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/0oOet2f43PA68X5RxKobEy"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/0oOet2f43PA68X5RxKobEy",
-                        "id": "0oOet2f43PA68X5RxKobEy",
-                        "name": "Shreya Ghoshal",
-                        "type": "artist",
-                        "uri": "spotify:artist:0oOet2f43PA68X5RxKobEy"
-                    }
-                ],
-                "available_markets": [],
-                "disc_number": 1,
-                "duration_ms": 263497,
-                "explicit": false,
-                "external_ids": {
-                    "isrc": "INZ031403751"
-                },
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/track/7xXhb2WxLJ0zIiEyxFVz31"
-                },
-                "href": "https://api.spotify.com/v1/tracks/7xXhb2WxLJ0zIiEyxFVz31",
-                "id": "7xXhb2WxLJ0zIiEyxFVz31",
-                "is_local": false,
-                "name": "Soona Man Ka Aangan",
-                "popularity": 31,
-                "preview_url": null,
-                "track_number": 3,
-                "type": "track",
-                "uri": "spotify:track:7xXhb2WxLJ0zIiEyxFVz31"
-            },
-            {
-                "album": {
-                    "album_type": "single",
-                    "artists": [
-                        {
-                            "external_urls": {
-                                "spotify": "https://open.spotify.com/artist/5bPD1UQmrfBLP450lSDkq4"
-                            },
-                            "href": "https://api.spotify.com/v1/artists/5bPD1UQmrfBLP450lSDkq4",
-                            "id": "5bPD1UQmrfBLP450lSDkq4",
-                            "name": "Sonu Sebastian",
-                            "type": "artist",
-                            "uri": "spotify:artist:5bPD1UQmrfBLP450lSDkq4"
-                        },
-                        {
-                            "external_urls": {
-                                "spotify": "https://open.spotify.com/artist/7IrZpbKw6j5L70Ymxr6ysg"
-                            },
-                            "href": "https://api.spotify.com/v1/artists/7IrZpbKw6j5L70Ymxr6ysg",
-                            "id": "7IrZpbKw6j5L70Ymxr6ysg",
-                            "name": "Shamil",
-                            "type": "artist",
-                            "uri": "spotify:artist:7IrZpbKw6j5L70Ymxr6ysg"
-                        }
-                    ],
-                    "available_markets": [
-                    ],
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/album/5dZgR2bdjBjD2lbYRx7x5n"
-                    },
-                    "href": "https://api.spotify.com/v1/albums/5dZgR2bdjBjD2lbYRx7x5n",
-                    "id": "5dZgR2bdjBjD2lbYRx7x5n",
-                    "images": [
-                        {
-                            "height": 640,
-                            "url": "https://i.scdn.co/image/ab67616d0000b273ed9a851a86b91ca825e0c9d1",
-                            "width": 640
-                        },
-                        {
-                            "height": 300,
-                            "url": "https://i.scdn.co/image/ab67616d00001e02ed9a851a86b91ca825e0c9d1",
-                            "width": 300
-                        },
-                        {
-                            "height": 64,
-                            "url": "https://i.scdn.co/image/ab67616d00004851ed9a851a86b91ca825e0c9d1",
-                            "width": 64
-                        }
-                    ],
-                    "name": "Splash",
-                    "release_date": "2023-10-27",
-                    "release_date_precision": "day",
-                    "total_tracks": 2,
-                    "type": "album",
-                    "uri": "spotify:album:5dZgR2bdjBjD2lbYRx7x5n"
-                },
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/5bPD1UQmrfBLP450lSDkq4"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/5bPD1UQmrfBLP450lSDkq4",
-                        "id": "5bPD1UQmrfBLP450lSDkq4",
-                        "name": "Sonu Sebastian",
-                        "type": "artist",
-                        "uri": "spotify:artist:5bPD1UQmrfBLP450lSDkq4"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/7IrZpbKw6j5L70Ymxr6ysg"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/7IrZpbKw6j5L70Ymxr6ysg",
-                        "id": "7IrZpbKw6j5L70Ymxr6ysg",
-                        "name": "Shamil",
-                        "type": "artist",
-                        "uri": "spotify:artist:7IrZpbKw6j5L70Ymxr6ysg"
-                    }
-                ],
-                "available_markets": [],
-                "disc_number": 1,
-                "duration_ms": 138461,
-                "explicit": false,
-                "external_ids": {
-                    "isrc": "ATEK02355908"
-                },
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/track/6KQTArsron0FgOOdSqZ8hd"
-                },
-                "href": "https://api.spotify.com/v1/tracks/6KQTArsron0FgOOdSqZ8hd",
-                "id": "6KQTArsron0FgOOdSqZ8hd",
-                "is_local": false,
-                "name": "Splash",
-                "popularity": 38,
-                "preview_url": "https://p.scdn.co/mp3-preview/fddf4c0355abb0a5aa16746a2da4ad1821c8d638?cid=f465a7e5d85849a182e92671b03b8973",
-                "track_number": 1,
-                "type": "track",
-                "uri": "spotify:track:6KQTArsron0FgOOdSqZ8hd"
-            },
-            {
-                "album": {
-                    "album_type": "single",
-                    "artists": [
-                        {
-                            "external_urls": {
-                                "spotify": "https://open.spotify.com/artist/7AKO4PWJ6ToZVUfA5xmsIY"
-                            },
-                            "href": "https://api.spotify.com/v1/artists/7AKO4PWJ6ToZVUfA5xmsIY",
-                            "id": "7AKO4PWJ6ToZVUfA5xmsIY",
-                            "name": "Sunanda Sharma",
-                            "type": "artist",
-                            "uri": "spotify:artist:7AKO4PWJ6ToZVUfA5xmsIY"
-                        }
-                    ],
-                    "available_markets": [],
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/album/0JTBanRxutQ7j4bRNgjh40"
-                    },
-                    "href": "https://api.spotify.com/v1/albums/0JTBanRxutQ7j4bRNgjh40",
-                    "id": "0JTBanRxutQ7j4bRNgjh40",
-                    "images": [
-                        {
-                            "height": 640,
-                            "url": "https://i.scdn.co/image/ab67616d0000b27358e1cff58934490d0bc4ba34",
-                            "width": 640
-                        },
-                        {
-                            "height": 300,
-                            "url": "https://i.scdn.co/image/ab67616d00001e0258e1cff58934490d0bc4ba34",
-                            "width": 300
-                        },
-                        {
-                            "height": 64,
-                            "url": "https://i.scdn.co/image/ab67616d0000485158e1cff58934490d0bc4ba34",
-                            "width": 64
-                        }
-                    ],
-                    "name": "Pagal Nahi Hona",
-                    "release_date": "2021-01-15",
-                    "release_date_precision": "day",
-                    "total_tracks": 1,
-                    "type": "album",
-                    "uri": "spotify:album:0JTBanRxutQ7j4bRNgjh40"
-                },
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/7AKO4PWJ6ToZVUfA5xmsIY"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/7AKO4PWJ6ToZVUfA5xmsIY",
-                        "id": "7AKO4PWJ6ToZVUfA5xmsIY",
-                        "name": "Sunanda Sharma",
-                        "type": "artist",
-                        "uri": "spotify:artist:7AKO4PWJ6ToZVUfA5xmsIY"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/1RViT01v3KsUvaEQk136nY"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/1RViT01v3KsUvaEQk136nY",
-                        "id": "1RViT01v3KsUvaEQk136nY",
-                        "name": "Sonu Sood",
-                        "type": "artist",
-                        "uri": "spotify:artist:1RViT01v3KsUvaEQk136nY"
-                    }
-                ],
-                "available_markets": [],
-                "disc_number": 1,
-                "duration_ms": 226046,
-                "explicit": false,
-                "external_ids": {
-                    "isrc": "INM231900507"
-                },
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/track/1BPnrHYkrXc28xp66lB2CZ"
-                },
-                "href": "https://api.spotify.com/v1/tracks/1BPnrHYkrXc28xp66lB2CZ",
-                "id": "1BPnrHYkrXc28xp66lB2CZ",
-                "is_local": false,
-                "name": "Pagal Nahi Hona",
-                "popularity": 43,
-                "preview_url": "https://p.scdn.co/mp3-preview/a7428a42097b89b70af8dfa564123864248bf781?cid=f465a7e5d85849a182e92671b03b8973",
-                "track_number": 1,
-                "type": "track",
-                "uri": "spotify:track:1BPnrHYkrXc28xp66lB2CZ"
-            },
-            {
-                "album": {
-                    "album_type": "album",
-                    "artists": [
-                        {
-                            "external_urls": {
-                                "spotify": "https://open.spotify.com/artist/4YgUVg4p7xtMOrOS4GjiJZ"
-                            },
-                            "href": "https://api.spotify.com/v1/artists/4YgUVg4p7xtMOrOS4GjiJZ",
-                            "id": "4YgUVg4p7xtMOrOS4GjiJZ",
-                            "name": "Jatin-Lalit",
-                            "type": "artist",
-                            "uri": "spotify:artist:4YgUVg4p7xtMOrOS4GjiJZ"
-                        }
-                    ],
-                    "available_markets": [],
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/album/2HeqbIcBcggDsgNCHcGN1h"
-                    },
-                    "href": "https://api.spotify.com/v1/albums/2HeqbIcBcggDsgNCHcGN1h",
-                    "id": "2HeqbIcBcggDsgNCHcGN1h",
-                    "images": [
-                        {
-                            "height": 640,
-                            "url": "https://i.scdn.co/image/ab67616d0000b27305a0d368c0fbf69c9a76bc81",
-                            "width": 640
-                        },
-                        {
-                            "height": 300,
-                            "url": "https://i.scdn.co/image/ab67616d00001e0205a0d368c0fbf69c9a76bc81",
-                            "width": 300
-                        },
-                        {
-                            "height": 64,
-                            "url": "https://i.scdn.co/image/ab67616d0000485105a0d368c0fbf69c9a76bc81",
-                            "width": 64
-                        }
-                    ],
-                    "name": "Mohabbatein",
-                    "release_date": "2000-08-27",
-                    "release_date_precision": "day",
-                    "total_tracks": 9,
-                    "type": "album",
-                    "uri": "spotify:album:2HeqbIcBcggDsgNCHcGN1h"
-                },
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/4YgUVg4p7xtMOrOS4GjiJZ"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/4YgUVg4p7xtMOrOS4GjiJZ",
-                        "id": "4YgUVg4p7xtMOrOS4GjiJZ",
-                        "name": "Jatin-Lalit",
-                        "type": "artist",
-                        "uri": "spotify:artist:4YgUVg4p7xtMOrOS4GjiJZ"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/70B80Lwx2sxti0M1Ng9e8K"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/70B80Lwx2sxti0M1Ng9e8K",
-                        "id": "70B80Lwx2sxti0M1Ng9e8K",
-                        "name": "Udit Narayan",
-                        "type": "artist",
-                        "uri": "spotify:artist:70B80Lwx2sxti0M1Ng9e8K"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/4qf5iWCSqeCW9TlbwO58bo"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/4qf5iWCSqeCW9TlbwO58bo",
-                        "id": "4qf5iWCSqeCW9TlbwO58bo",
-                        "name": "Jaspinder Narula",
-                        "type": "artist",
-                        "uri": "spotify:artist:4qf5iWCSqeCW9TlbwO58bo"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/6lJSXbir8V0GruDu5r07Un"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/6lJSXbir8V0GruDu5r07Un",
-                        "id": "6lJSXbir8V0GruDu5r07Un",
-                        "name": "Udbhav",
-                        "type": "artist",
-                        "uri": "spotify:artist:6lJSXbir8V0GruDu5r07Un"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/4OYgq5HqicvfcYOIyjsvYE"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/4OYgq5HqicvfcYOIyjsvYE",
-                        "id": "4OYgq5HqicvfcYOIyjsvYE",
-                        "name": "Manohar Shetty",
-                        "type": "artist",
-                        "uri": "spotify:artist:4OYgq5HqicvfcYOIyjsvYE"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/1UArumT3RYiSSbvYTxXRAY"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/1UArumT3RYiSSbvYTxXRAY",
-                        "id": "1UArumT3RYiSSbvYTxXRAY",
-                        "name": "Ishaan",
-                        "type": "artist",
-                        "uri": "spotify:artist:1UArumT3RYiSSbvYTxXRAY"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/4L0JycQCCwKksT3G5SVApu"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/4L0JycQCCwKksT3G5SVApu",
-                        "id": "4L0JycQCCwKksT3G5SVApu",
-                        "name": "Shweta Pandit",
-                        "type": "artist",
-                        "uri": "spotify:artist:4L0JycQCCwKksT3G5SVApu"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/3m1AeR9ODWpnXqeHQ4VdS7"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/3m1AeR9ODWpnXqeHQ4VdS7",
-                        "id": "3m1AeR9ODWpnXqeHQ4VdS7",
-                        "name": "Sonali Bhatawdekar",
-                        "type": "artist",
-                        "uri": "spotify:artist:3m1AeR9ODWpnXqeHQ4VdS7"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/0NHOjXrjKEiChop2sZkhdS"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/0NHOjXrjKEiChop2sZkhdS",
-                        "id": "0NHOjXrjKEiChop2sZkhdS",
-                        "name": "Pritha Mazumdar",
-                        "type": "artist",
-                        "uri": "spotify:artist:0NHOjXrjKEiChop2sZkhdS"
-                    },
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/3Bku4b0uiR8KJSXu7U8Iyy"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/3Bku4b0uiR8KJSXu7U8Iyy",
-                        "id": "3Bku4b0uiR8KJSXu7U8Iyy",
-                        "name": "Anand Bakshi",
-                        "type": "artist",
-                        "uri": "spotify:artist:3Bku4b0uiR8KJSXu7U8Iyy"
-                    }
-                ],
-                "available_markets": [],
-                "disc_number": 1,
-                "duration_ms": 547422,
-                "explicit": false,
-                "external_ids": {
-                    "isrc": "INY090000005"
-                },
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/track/5lLBVDoG0w9oVxEftMIniN"
-                },
-                "href": "https://api.spotify.com/v1/tracks/5lLBVDoG0w9oVxEftMIniN",
-                "id": "5lLBVDoG0w9oVxEftMIniN",
-                "is_local": false,
-                "name": "Soni Soni",
-                "popularity": 55,
-                "preview_url": "https://p.scdn.co/mp3-preview/385468aa70e31ff4d2e50d07bfe361d16b22946f?cid=f465a7e5d85849a182e92671b03b8973",
-                "track_number": 5,
-                "type": "track",
-                "uri": "spotify:track:5lLBVDoG0w9oVxEftMIniN"
-            },
-            {
-                "album": {
-                    "album_type": "album",
-                    "artists": [
-                        {
-                            "external_urls": {
-                                "spotify": "https://open.spotify.com/artist/0y633xrd3L3rSxpZ4awgnG"
-                            },
-                            "href": "https://api.spotify.com/v1/artists/0y633xrd3L3rSxpZ4awgnG",
-                            "id": "0y633xrd3L3rSxpZ4awgnG",
-                            "name": "Ravindra Jain",
-                            "type": "artist",
-                            "uri": "spotify:artist:0y633xrd3L3rSxpZ4awgnG"
-                        }
-                    ],
-                    "available_markets": [],
-                    "external_urls": {
-                        "spotify": "https://open.spotify.com/album/1rYG8BTCZeZC7uJYZnsfcm"
-                    },
-                    "href": "https://api.spotify.com/v1/albums/1rYG8BTCZeZC7uJYZnsfcm",
-                    "id": "1rYG8BTCZeZC7uJYZnsfcm",
-                    "images": [
-                        {
-                            "height": 640,
-                            "url": "https://i.scdn.co/image/ab67616d0000b2736417d0deffec0833dfabec11",
-                            "width": 640
-                        },
-                        {
-                            "height": 300,
-                            "url": "https://i.scdn.co/image/ab67616d00001e026417d0deffec0833dfabec11",
-                            "width": 300
-                        },
-                        {
-                            "height": 64,
-                            "url": "https://i.scdn.co/image/ab67616d000048516417d0deffec0833dfabec11",
-                            "width": 64
-                        }
-                    ],
-                    "name": "Ek Vivaah..Aisa Bhi",
-                    "release_date": "2009-01-01",
-                    "release_date_precision": "day",
-                    "total_tracks": 12,
-                    "type": "album",
-                    "uri": "spotify:album:1rYG8BTCZeZC7uJYZnsfcm"
-                },
-                "artists": [
-                    {
-                        "external_urls": {
-                            "spotify": "https://open.spotify.com/artist/0w4e7HVbqAHFPN6VkIlvhe"
-                        },
-                        "href": "https://api.spotify.com/v1/artists/0w4e7HVbqAHFPN6VkIlvhe",
-                        "id": "0w4e7HVbqAHFPN6VkIlvhe",
-                        "name": "Suresh Wadkar",
-                        "type": "artist",
-                        "uri": "spotify:artist:0w4e7HVbqAHFPN6VkIlvhe"
-                    }
-                ],
-                "available_markets": [],
-                "disc_number": 1,
-                "duration_ms": 155632,
-                "explicit": false,
-                "external_ids": {
-                    "isrc": "INS180907991"
-                },
-                "external_urls": {
-                    "spotify": "https://open.spotify.com/track/4MhP0ZvD12YMMVcHSHrqo4"
-                },
-                "href": "https://api.spotify.com/v1/tracks/4MhP0ZvD12YMMVcHSHrqo4",
-                "id": "4MhP0ZvD12YMMVcHSHrqo4",
-                "is_local": false,
-                "name": "Mandir Hai Ghar Ye Hamara",
-                "popularity": 26,
-                "preview_url": "https://p.scdn.co/mp3-preview/b5957a0f4d214e300731320efa7ec51f0a32d13d?cid=f465a7e5d85849a182e92671b03b8973",
-                "track_number": 4,
-                "type": "track",
-                "uri": "spotify:track:4MhP0ZvD12YMMVcHSHrqo4"
-            }
-        ],
-        "limit": 5,
-        "next": "https://api.spotify.com/v1/search?query=sonu+soo&type=track&locale=en-GB%2Cen-US%3Bq%3D0.9%2Cen%3Bq%3D0.8%2Chi%3Bq%3D0.7&offset=5&limit=5",
-        "offset": 0,
-        "previous": null,
-        "total": 813
-    },
-}
