--- src/components/HeaderSearch/index.ts
@@ -1,1 +1,0 @@
-export { HeaderSearch } from "./HeaderSearch"
