--- src/pages/SearchResultPage.tsx
@@ -1,9 +1,0 @@
-
-export const SearchResult = () => {
-
-    return (
-        <div>
-            
-        </div>
-    )
-}
